updating with events and asset queue
