+++ engine/includes/Objects/Assets/AssetQueue.hpp
@@ -1,0 +1,25 @@
+class AssetQueue { // extends Queue?
+/*
+each scene will have a scene file with
+dependent assets
+
+on scene load, pull assets
+
+should scenes be partitioned by user or
+by engine?
+
+should users be incentivised to optimize
+asset loading or do we do the heavy lifting?
+
+*/
+
+public:
+	/*
+	load asset breakdown : file paths
+	for i in assets:
+		...
+	*/
+
+private:
+
+}
