+++ engine/includes/Scene/Scene.hpp
@@ -1,0 +1,9 @@
+class Scene {
+public:
+
+private:
+	AssetQueue *asset_queue;
+	
+	std::vector<Vertex> vertices;
+	std::vector<uint32_t> indices;
+}
