+++ engine/includes/Scene/SceneManager.hpp
@@ -1,0 +1,7 @@
+class SceneManager {
+public:
+
+private:
+	// scene graph
+	// try to use boost++
+}
