+++ engine/includes/render/RenderMgr.hpp
@@ -11,11 +11,6 @@
 #include <X11/Xlib.h>
 #include <chrono>
 
-#ifndef GLFW_INCLUDE_VULKAN
-#define GLFW_INCLUDE_VULKAN
-#include <GLFW/glfw3.h>
-#endif
-
 #define GLM_ENABLE_EXPERIMENTAL
 #include <glm/gtx/hash.hpp>
 
@@ -24,28 +19,7 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 
-#ifndef SCRNWOFFS
-#define SCRNWOFFS -100
-#endif
-
-#ifndef SCRNHOFFS
-#define SCRNHOFFS - 100
-#endif
-
-const std::string MODEL_PATH = "/home/samurai/dark-engine/engine/source/render/models/chalet.obj";
-const std::string TEXTURE_PATH = "/home/samurai/dark-engine/engine/source/render/textures/chalet.jpg";
-
-#ifndef TEST_VERT_FILE
-#define TEST_VERT_FILE "/home/samurai/dark-engine/engine/source/render/shaders/vert.spv"
-#endif
-
-#ifndef TEST_FRAG_FILE
-#define TEST_FRAG_FILE "/home/samurai/dark-engine/engine/source/render/shaders/frag.spv"
-#endif
-
 const int MAX_FRAMES_IN_FLIGHT = 2;
-
-#include <GLFW/glfw3.h>
 
 const std::vector<const char*> validationLayers = {
     "VK_LAYER_LUNARG_standard_validation"
@@ -87,7 +61,7 @@ struct	QueueFamilyIndices {
 };
 
 /*
-TODO : move this shit into another file ok
+TODO : move this into another file ok
 Vertex.h/cpp
 */
 #include <glm/glm.hpp>
@@ -96,47 +70,6 @@ struct UniformBufferObject {
     alignas(16) glm::mat4 model;
     alignas(16) glm::mat4 view;
     alignas(16) glm::mat4 proj;
-};
-
-struct Vertex {
-	glm::vec3 pos;
-	glm::vec3 color;
-	glm::vec2 texCoord;
-
-	static VkVertexInputBindingDescription getBindingDescription() {
-		// BIG TODO : needs reworking dood
-		VkVertexInputBindingDescription bindingDescription = {};
-		bindingDescription.binding = 0;
-		bindingDescription.stride = sizeof(Vertex);
-		bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-
-		return (bindingDescription);
-	}
-
-	static std::array<VkVertexInputAttributeDescription, 3> getAttributeDescriptions() {
-		std::array<VkVertexInputAttributeDescription, 3> attributeDescriptions = {};
-
-		attributeDescriptions[0].binding = 0;
-		attributeDescriptions[0].location = 0;
-        attributeDescriptions[0].format = VK_FORMAT_R32G32B32_SFLOAT;
-   		attributeDescriptions[0].offset = offsetof(Vertex, pos);
-
-		attributeDescriptions[1].binding = 0;
-        attributeDescriptions[1].location = 1;
-        attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[1].offset = offsetof(Vertex, color);
-
-		attributeDescriptions[2].binding = 0;
-		attributeDescriptions[2].location = 2;
-		attributeDescriptions[2].format = VK_FORMAT_R32G32_SFLOAT;
-        attributeDescriptions[2].offset = offsetof(Vertex, texCoord);
-
-		return attributeDescriptions;
-	}
-	
-    bool operator==(const Vertex& other) const {
-        return pos == other.pos && color == other.color && texCoord == other.texCoord;
-    }
 };
 
 namespace std {
@@ -179,17 +112,8 @@ public:
 	/* Getters */
 	bool									getScreenRes();
 	// TODO : preprocessor statements for debugging
-	uint32_t								getScreenWidth() {
-			// return (800);}
-		return (_screen->width + SCRNWOFFS);}
-	uint32_t								getScreenHeight() {
-			// return (400);}
-		return (_screen->height + SCRNHOFFS);}
 
 	/* Setters */
-	void									setFrameBufferResized(bool stat) {
-		framebufferResized = stat;
-	}
 
 private:
 	// TODO : move to utils.cpp
@@ -203,7 +127,7 @@ private:
 	bool									createFrameBuffers();
 	bool									createCommandPool();
 	bool									createCommandBuffers();
-void endSingleTimeCommands(VkCommandBuffer commandBuffer);
+	void endSingleTimeCommands(VkCommandBuffer commandBuffer);
 
 	bool									createSurface();
 	bool									createSwapChain();
@@ -215,14 +139,13 @@ void endSingleTimeCommands(VkCommandBuffer commandBuffer);
 	VkExtent2D								chooseSwapExtent(const VkSurfaceCapabilitiesKHR& capabilities);
 	bool									createImageViews();
 	void									createSemaphores();
-	static void								framebufferResizeCallback(GLFWwindow* window, int width, int height);
 	bool createUniformBuffers();
 	bool createDescriptorPool();
 	void updateUniformBuffer(uint32_t currentImage);
 	VkShaderModule							createShaderModule(const std::vector<char>& code);
 	bool createDescriptorSetLayout();
 	bool createDescriptorSets();
-VkSampleCountFlagBits getMaxUsableSampleCount();
+	VkSampleCountFlagBits getMaxUsableSampleCount();
 	bool 									createIndexBuffer();
 	bool									pickPhysicalDevice();
 	int										rateDeviceSuitability(VkPhysicalDevice device);
@@ -236,7 +159,7 @@ VkSampleCountFlagBits getMaxUsableSampleCount();
 	bool 									createSyncObjects();
 	bool									setupDebugCallback();
 	bool 									createVertexBuffer();
-VkFormat findSupportedFormat(const std::vector<VkFormat>& candidates, VkImageTiling tiling, VkFormatFeatureFlags features);
+	VkFormat findSupportedFormat(const std::vector<VkFormat>& candidates, VkImageTiling tiling, VkFormatFeatureFlags features);
 	bool createDepthResources();
 	void transitionImageLayout(VkImage image, VkFormat format, VkImageLayout oldLayout, VkImageLayout newLayout, uint32_t mipLevels);
 	void copyBufferToImage(VkBuffer buffer, VkImage image, uint32_t width, uint32_t height);
@@ -274,13 +197,7 @@ VkFormat findSupportedFormat(const std::vector<VkFormat>& candidates, VkImageTil
 
 	/* Platform specs */
 //	RECT						_desktop;
-	Display								*_display;
-	Screen								*_screen;
-	int									_width;
-	int									_height;
-
-	/* GLFW specs */
-	GLFWwindow* 						_window;
+	DarkEngine 							*dark_engine;
 
     VkInstance							_instance;
     VkDebugUtilsMessengerEXT 			callback;
@@ -290,8 +207,8 @@ VkFormat findSupportedFormat(const std::vector<VkFormat>& candidates, VkImageTil
     VkDevice 							_device;
     VkQueue 							graphicsQueue;
     VkImage depthImage;
-VkDeviceMemory depthImageMemory;
-VkImageView depthImageView;
+	VkDeviceMemory depthImageMemory;
+	VkImageView depthImageView;
     VkQueue 							presentQueue;
 	VkDescriptorPool descriptorPool;
 	std::vector<VkDescriptorSet> descriptorSets;
@@ -305,13 +222,12 @@ VkImageView depthImageView;
     std::vector<VkImageView> 			swapChainImageViews;
     std::vector<VkFramebuffer> 			swapChainFramebuffers;
     VkImage colorImage;
-VkDeviceMemory colorImageMemory;
-VkImageView colorImageView;
+	VkDeviceMemory colorImageMemory;
+	VkImageView colorImageView;
     VkImageView textureImageView;
     uint32_t mipLevels;
 	VkSampler textureSampler;
-	std::vector<Vertex> vertices;
-	std::vector<uint32_t> indices;
+
 	VkDeviceMemory vertexBufferMemory;
 	std::vector<VkBuffer> uniformBuffers;
 	std::vector<VkDeviceMemory> uniformBuffersMemory;
