+++ engine/source/objects/Assets/Asset.cpp
@@ -1,0 +1,1 @@
+// Asset.cpp
