+++ engine/source/objects/Assets/AssetQueue.cpp
@@ -1,0 +1,69 @@
+// AssetQueue.cpp
+#ifndef ASSETQUEUE_H
+#define ASSETQUEUE_H
+
+const std::string MODEL_PATH = "/home/samurai/dark-engine/engine/source/render/models/chalet.obj";
+const std::string TEXTURE_PATH = "/home/samurai/dark-engine/engine/source/render/textures/chalet.jpg";
+
+#ifndef TEST_VERT_FILE
+#define TEST_VERT_FILE "/home/samurai/dark-engine/engine/source/render/shaders/vert.spv"
+#endif
+
+#ifndef TEST_FRAG_FILE
+#define TEST_FRAG_FILE "/home/samurai/dark-engine/engine/source/render/shaders/frag.spv"
+#endif
+
+void AssetQueue::load_model() {
+	/*
+	parallize whatever parent is calling this
+	*/
+
+    tinyobj::attrib_t attrib;
+    std::vector<tinyobj::shape_t> shapes;
+    std::vector<tinyobj::material_t> materials;
+    std::string warn, err;
+
+    if (!tinyobj::LoadObj(
+    	&attrib, &shapes, &materials,
+    	&warn, &err, MODEL_PATH.c_str())
+	) {
+        throw std::runtime_error(warn + err);
+    }
+
+	std::unordered_map<Vertex, uint32_t> uniqueVertices = {};
+	for (const auto& shape : shapes) {
+	    for (const auto& index : shape.mesh.indices) {
+	        Vertex vertex = {};
+	        
+            vertex.pos = {
+                attrib.vertices[3 * index.vertex_index + 0],
+                attrib.vertices[3 * index.vertex_index + 1],
+                attrib.vertices[3 * index.vertex_index + 2]
+            };
+
+            vertex.texCoord = {
+                attrib.texcoords[2 * index.texcoord_index + 0],
+                1.0f - attrib.texcoords[2 * index.texcoord_index + 1]
+            };
+
+            vertex.color = {1.0f, 1.0f, 1.0f};
+
+            // only load the vertices once
+	        if (uniqueVertices.count(vertex) == 0) {
+	            uniqueVertices[vertex] = static_cast<uint32_t>(vertices.size());
+	            vertices.push_back(vertex);
+	        }
+
+	        indices.push_back(uniqueVertices[vertex]);
+	    }
+	}
+}
+
+void AssetQueue::loadModels() {
+	/*
+	for i in paths:
+		load_model(path);
+	*/
+}
+
+#endif // ASSETQUEUE_H
